Turn lines into line views, react to things.
Wiring, wiring, wiring. Also thinking about how things will be intertwingled.
As usual, more than I'd prefer. I don't like that LineViewImpl creates the node
in the absence of knowledge of the CodeArea, but maybe whatever the CodeArea
would tell it, it can get from some more fundamental source, and that will be
good. Otherwise we end up with a Gordian knot of recursive dependencies between
BufferViewImpl, KeyDispatcher, the MajorMode and CodeArea, which I don't like.