More key dispatch revampery.
I'm slowly grokking JavaFX key dispatch. It's still not ideal (I still can't
support e.g. C-_ directly) but it is looking less boneheaded. :)

Modified key presses (via alt, ctrl, shift) are translated by the OS into
alternate characters. In the case of ALT, that ends up being Greeky characters,
in the case of SHIFT, it's upper-case and shifted characters (natch), and in
the case of CTRL it's VT100 control characters (for example, CTRL-g is ASCII
7). The CTRL key map is a little wonky because not every keyboard key has a
control equivalent. So CTRL-= is just = (ASCII 61).

The still boneheaded thing is that if you use SHIFT in conjunction with any
other modifier key, it reverts to reporting the non-shifted character (e.g.
SHIFT-= gives + but CTRL-SHIFT-= gives =). This is non-awesome because I'd like
you to be able to bind C-+ directly without having to assume that the keyboard
in question has + above = and bind C-S-=. I need to file a JavaFX bug and see
what can be done.