Various "shell" thinky and revampy.
We're going to call a single "window" an Editor and it will contain one or more
tabs each of which contains a BufferArea displaying a buffer for editing. The
kill ring and minibuffer will be shared by the whole editor because that's a
nice level of grouping.

It will eventually be possible to have multiple editors open (one for each
virtual console, say) and each will largely be its own independent domain. The
buffer list, kill ring, minibuffer history, etc. for one editor will not
intertwingle with the list for another. Multiple editors may share services
(i.e. "scala compiler service", once those are implemented), since services
will tend to be ignorant of buffers and views and whatnot and there's no point
in being unnecessarily profligate with memory. The main point of supporting
multiple editors in the same process is to avoid the bloat of running multiple
separate scaled processes.