Minibuffer revamp; no more read stack.
We now have one active read and any further attempts to read are rejected.
Status can "stack" over an active read and when the status goes away we return
to the read, but otherwise you just do one thing at a time.

Emacs allows limited read stacking, but I've always found that confusing so I'm
not champing at the bit to recreate that functionality. We'll see...