Move undo stack from view to buffer.
Even though undoing manipulates the point of the view that undoes the buffer,
Emacs uses the same undo stack for all views of a buffer (which makes sense).
So we pass the point in from the view that's making changes and then restore
that point to whatever view undoes/redoes the changes.