Out with f*ing JavaFX tabs.
The focus issues were super annoying and it's not really that useful to see
what buffers you have open across the top of the view. If you're even remotely
hardcore there will be too many anyway.

Eventually I'd like to deemphasize the notion of "open" buffers anyway, and
move to a "go to this file in this project" intent, which won't care if the
file is open yet or not.