Some special handling for non-existent files.
We still want as "canonical" as possible a path even for files that don't
exist. So if you edit "foo/bar" we want to prepend the cwd before passing it
on. As the comment says, it would be nice if the caller didn't have to do this,
but once we turn it into a file this becomes fiddly.

This is a stopgap solution and I'll revisit at some point. I may yet ditch File
entirely and use NIO Path...