Revamped faces to be CSS styles.
This offloads a lot of the fiddling to JavaFX's CSS implementation, which is
great, but I will need to do some higher level shenanigans to support
"inherited" faces.

For example, Emacs's scala-mode2 defines a variable face that inherits from
warning face, so if you don't customize it, it inherits whatever style warning
face gets. That's nice because if you "theme" your editor, warning face gets
changed to something by the theme and Scala's vars magically use that theme.

AFAIK, we can't inherit CSS styles in this way, so I'll probably have to
reintroduce Face as a wrapper around a CSS style along with an optional parent
face. Then when I apply the face, I'll check to see if the CSS style is
actually defined in any of the style sheets, and if not, fall back to the
parent face (and so forth).

This whole business was also substantially complicated by the fact that there's
no way to fill in the background rectangle of a Text node. So I had to create a
FillableText node which exports a new -fx-bg-fill style and which maintains a
Rectangle with the bg fill. The passages are always twisty and maze-like.