Added Behavior. Improved whitespace mode.
Behavior makes it easy to encapsulate some functionality that wires up a bunch
of reactions when enabled, and then clears them all out when disabled. This can
then be bound to a boolean config property so that things are automatically
wired and unwired if the user changes the property interactively.

WhitespaceMode now adds and removes the trailing whitespace style, as
appropriate. It doesn't highlight any trailing whitepace to the left of the
point. It defers trailing whitespace processing a frame so that edits and point
movements are all applied before it takes a look at things.