More minibuffer tweaking.
We now provide a built-in completion display because we can include it in the
minibuffer overlay nicely. There is still much to be done to make things behave
sensibly when completing long things or with lots of completions.

The jury is still also out on multiple columns in the completions display. I'm
thinking I probably do want them...