Buffer/View properties now use Property.
This allows us to expose non-reactive versions in the Foo classes and then
simultaneously refine and implement those with reactive versions in the RFoo
class.

It is unfortunate that callers have to sprinkle () all over the place, but I
prefer not to repeat myself four times just to declare a property, and I like
that this unifies everything under a single name. No more foo and fooV.