Some layout related jiggery pokery.
There is still an annoying JavaFX "bug" wherein if the first minibuffer
invocation uses a label with spaces in it (i.e. which could be wrapped) the
label prefers a bizarrely tall height (like 80 instead of 30) which then causes
the initial minibuffer to be 4x taller than it should be because the label
somehow sorts its shit out before actual layout happens and the minibuffer is
left holding all the extra space because the BorderArea has already committed
to the extra tall layout and the minibuffer is in the center, where extra space
is placed.

I've already sunk a few hours trying to trace through the twisty fucking maze
of insanity that is JavaFX's "content biased" layout code and have not yet
figured out where the major malfunction is. So we'll just live with this wart
for now. Sigh.