Added MajorMode.missedFn, wired up disposal.
We're going to need missedFn to allow ISearchMode to reroute most key bindings
back to the major mode that triggered it. When it sees a non-isearch key
sequence, it will terminate the search and pass the key sequence back to the
invoking major mode for execution.

DispatcherImpl.dispose() will be used to ensure that modes are properly
disposed when a BufferArea is going away. This is also chiefly useful for
minibuffer modes where they may want to clean up on disposal.