Add the ability to "execute" a key press.
You pass in the text description of a trigger sequence, like C-x C-c and it
parses it, then resolves the fn bound thereto, and invokes it. If there's no fn
bound, it invokes the standard missedFn processing so that the same feedback
occurs as if you typed the unbound key sequence normally.