Wired up loose case matching.
If you search for an all lowercase string, either case is matched. If you
search for a mixed case (or all uppercase) string, exact match is required.

This has got to be one of the best editor inventions ever, I use it in
everything I write that has searching. I sure am glad it's not patented.