More work on config, it grows to encompass vars.
We now have two levels of configuration: editor-wide, and mode-wide. Config
keys are tagged as global or not, and resolved in the appropriate ConfigImpl
based on that.

Config also acts as a shraed state storage. For example, the kill ring is now a
global config entry rather than a privileged field in the Editor trait. While
the user will probably not actually customize the kill ring, they could if they
were doing something ULTRA CRAZY.

Modes can share state between simultaneous or time separated instantiations by
using their mode-wide config. This is unsurprising in the case of
configuration, but it also makes sense for them to share mode-wide rings or
other mutable state therein.