Wired up scanning for annotated classes.
Right now we only scan the "built-in" package, but it will be trivial to start
scanning external packages once I sort out how the directory structure will
work there.

I kind of want just a single top-level directory for each package:

.../Packages/whitespace-mode
.../Packages/project-service
.../Packages/scaled-scala-mode

but the packages are actually identified by a version control URL, so it might
be better to have:

.../Packages/github.com/scaled/whitespace-mode
.../Packages/github.com/scaled/project-service
.../Packages/code.google.com/p/scaled-scala-mode

but if I do the latter, then I need some sort of marker file at the top-level
of each package. Now that I say this out loud, I realize that the latter
approach is clearly better and the marker file will be useful for other
reasons. Make it so!