Various work on package management.
Revamped the package.scaled file a bit to make the names more uniformly
verbose. Also wired up half-assed local Ivy depends. The other half of that ass
and Maven depends will come eventually, but this is enough to allow me to test
my own externally implemented major modes which themselves depend on the
separate textmate-grammar library (which is installed to ~/.ivy2).