Avoid highlighting single character matches.
It needlessly slows down the beginning of every search to highlight the zillion
instances of the first character and then unhighlight them as you add more
characters. If we eventually delay the display of non-active matches, this can
be removed, but I kind of prefer instant visualization of all matches.

I also added a count of the number of matches in the prompt when there are >0.