Move status to the lower quarter of the screen.
I feel like that overlaps with the text to which I'm attending less often. I
like having the minibuffer up top because one is naturally attending to the
minibuffer when it's open, but status is often incidental and ignorable and we
don't want to beat the user over the head with it.