Added Blocker for matching braces.
CodeMode wires this up to automatically highlight the braces that enclose the
tighest block around the point. I'll probably add a config to allow this to be
turned off, or only active when the point is on a brace, or blinked on request,
for people who don't like such enthusiastic block information.