Revamp the way we handle builtin packages.
Now everyone provides a package.scaled, even the main built-in package. We also
support running the editor in "everything in one classpath" mode, where the
editor and a bunch of packages are all rolled up into a single top-level
project and compiled and run all together.

This allows us to use JRebel, which chokes if we do the normal classpath
jockeying to make external packages work. It also allows us to have one SBT
build which rebuilds everything when anything changes.

This all might change when Scaled becomes completely self hosting.