Move max line length tracking into buffer.
BufferArea needs it, so we might as well make it available to everyone else
without having to track it themselves.

Also it's turning out to be fiddly to have max line length tracking happen as
part of the buffer edit notification chain. Other things can be earlier in that
chain than the max line length tracker and if those things ask the tracker for
the max line length, it could be stale.