Nixed SyntaxTable, added Chars.
We weren't really using the syntax table as Emacs did, and it's simpler to just
have some useful character functions and deal with syntax separately.

This also changes Buffer.findForward/Backward to take a three arg predicate
which is passed the row and column as well as the character, so that a
predicate can check other information about a character during matching.