Restrict block-matching to same syntax class.
This means brackets in code only match other brackets in code, which is good,
but also that brackets in comments only match other brackets in comments, and
same for strings.

I'm not 100% sure I like the syntax class approach versus just code/not-code,
but I'm going to roll with this for a while and see how I feel. It's a little
weird when working on code that "talks" about brackets a lot (like IDE code),
since that tends to have dangling brackets in code which look weird.