Don't tweak vertical scroll top unless we're way OOB.
Otherwise if you scroll to the bottom, then center the bottom line in the view,
any movement will scroll the bottom line down to the bottom view again. This is
annoying because you put the bottom line in the center of the view on purpose
and obviously want it to stay there.