Purge cr/lf from filenames used as completer text.
Sigh. I'm sure this won't be the last bandaid on this festering wound.

At least to Mac OS's credit, the Finder would not let me create a file with a
newline in the name to test this bug. I had to test it on a file created by a
Windows machine and then synced to my Dropbox.