Added replace-string, replace-regexp.
Also stubbed out query-replace. I need to chain a bunch of miniReadOpts, but
before I do that I want to decide how I'm going to handle recording
modifications like this to the undo stack.

Right now the replace modifications are not being properly recorded because
this action happens outside a fn invocation. Maybe I should just record
everything that happened since the start of the last fn at the beginning of the
next fn invocation. Or I could expose manual controls to flush edits to the
undo stack, but that seems like a recipe for future programmer error.