More project services bits, starting on compile API.
I'm leaning toward a bunch of special purpose interfaces for different aspects
of the "assisted coding" experience.

This first one is a batch-ish compiler interface that will be triggered on file
save and will be easiest to implement for not-so-tightly integrated build
systems, but will also work for fancier integrations, at higher speed.