Made mode resolution immediate.
It was actually immediate under the hood, I just had vestigial Future business
around. I'm going to stick with the new plan of making sure everything is
sorted during the package installation phase, so that when we get around to
resolving a mode, we know everything is in place and ready to go.