Bound the point into the buffer on read and write.
I probably need to address this more fundamentally, but it's troublesome when
the point becomes invalid because someone deletes a bunch of text from the
buffer, leaving the point dangling in space.

Most likely I need to just automatically adjust the point when additions and
deletions occur in the buffer prior to the point.