Revamped batch compilation support.
We're just going to go Emacs-style for this level of support. Most modes will
pipe the output of their compiler into a buffer, match file path and line
number from the output and for extra credit, parse out the error string so that
we can display it to the user when we move to the error in question.

The next level of sophistication will support tagging regions of the buffer
with faces to indicate warnings, errors, and whatnot, but that needs tighter
integration with the compiler than basic batch compilation provides.