(Re-)introducing Syntax.
Every character position now has a Syntax instance associated with it along
with a Styles instance. Syntax allows code to know basic syntax information
about the entire file without resorting to reverse engineering it from Styles
(which is error prone/lossy anyway).

Presently the only two things we care about are: is this a comment? and is this
a literal of some sort? This will be used by blocker and indenters to avoid
spuriously matching things inside comments, &c.

A mode can also use syntax for its own more sophisticated purposes, if desired,
as long as it properly expresses the isComment and isLiteral nature of it
special syntax instances.