Some temporary debug info.
Things start to slow down substantially after a while. It doesn't seem like a
memory leak, since we're not GCing a lot, but this debugging info also
indicates that it doesn't seem to be a ton of time spent in layoutChildren.

So I'm not sure what's baking who's noodle, but I'm leaving this
instrumentation in until it's all sorted.