Added invalidStyle.
This avoids the temptation to use warnStyle or errorStyle when styling syntax,
because that will cause problems when those two styles are not automatically
removed from restyled code.

I may have to rethink the style business (or the TextMate syntax to style
business) because that particular bear trap is pernicious.