Guard against on-save hooks preventing save.
I'm not thrilled with this construction, but the buffer doesn't have a log
(maybe it should have one), so better to save and rethrow the exception than
simply ignore it.

It's also not 100% pure awesome to blindly plow ahead with a save when one or
more save-hooks has choked, but it's less non-awesome than not being able to
save your file at all.