Extracted paragraph identification into helper.
This allows the helper to be more sophisticated about identifying paragraphs.
Specifically, in Javadoc comments we need to differentiate between the default
notion of delimiter (a blank line) and a line that itself is part of a
paragraph but should not be part of the paragraph above it (or vice versa).