Yet more revamping.
To avoid a bunch of annoying special casery, I'm going to have the JVM that's
running the package manager launch the editor directly. Then the editor will
just use the package manager as is to do its package stuffs.

I opted to make PackageManager a global singleton rather than pass it through
the editor's main() because the editor's main() is already greatly weirded by
the requirements of JavaFX Application. Meh.