Various revamping for new launcher approach.
We now magically inherit the PackageManager, so we can't customize it, we have
to observe it. Everything is almost wired up, except that I have to work out a
pesky niggle which is that scaled-api defines services that scaled-editor
implements, and the package manager is not set up to handle services being
implemented in a package other than the package in which they were defined.
Double meh.