Rewrote whole shebang in Java.
Boy that's annoying. We also now use Capsule (which in turn uses Aether) to
handle Maven dependencies, which means we're doing things "properly" and also
that Maven dependencies will be downloaded from Central automagically.

A bunch of work remains to be done, and I need to sort out how I'm going to
deal with the fact that everyone needs to share the same scala-library and
guava depends. Oh the funs.