A new new approach to transitive depends.
A package inherits all Maven depends from its project dependencies (and omits
them from its own list). This means it is not possible to override a
subpackage's dependency with a different version, but that's basically
impossible anyway because someone else may have already resolved that
subpackage and *not* overridden its depends and it will have loaded the
non-overridden classes, etc. etc.

So a package has to be lord and master over its own transitive depends. So when
you depend on a package, you get the "whole package" so to speak. Hopefully
this will not be a giant fucking quagmire down the road, but we'll see.