Implement system depends for tools.jar.
Annoying to have to introduce a whole new kind of dependency for this one
damned jar, but maybe system depends will turn out to be useful for other
things in the future.

I started implementing this with a special Maven identifier, but I couldn't
stand the hackery, so I went with the 5x-more-code approach.