Make completion more DWIM.
If the field was pre-populated with a valid completion, or you happened to just
type exactly what you wanted before ever pressing return, it was generating a
completion with that value and requiring you to press enter a second time to do
commit the action.

Now it sees that you're committing with no completion, generates a completion,
and commits immediately if what you've provided is a valid completion.