Moved searching out of Buffer, into Search.
This reorganizes things a bit as well to allow one to find the first match,
forward and backward, and all the matches, separately. This API is needed for
the (in-progress) i-search revamp.

The current bits seem like a rather simple wrapper around Buffer.findForward/
Backward, but when we eventually support multi-line search, the Search class
will carry its weight.