Added workspace mechanism to core editor.
I had this in the project service, but I think it's broadly useful enough that
it would be nice to have in the core editor. And that allows us to integrate
editors and workspaces a bit more tightly without having to do hackery or
provide a "single-use" public API.

I can also add "scaled -w workspace ..." without feeling bad about it.