Avoid Files.createDirectories.
It freaks out if one of the directories in the path happens to be a symlink
(regardless of whether that symlink is to a directory). This makes life
annoying when one is trying to share workspaces between editor installations.
This is chiefly only something someone working on Scaled would do, but hey,
that's me, and naturally I'm going to solve my own problems, however esoteric.