Open new workspaces on same desktop as needed.
If we have a hint path that indicates that a file belongs to a particular
workspace, then open that workspace, regardless of whether or not some other
workspace is already open on this desktop.

Right now a new editor will always be opened on the current desktop if no
editor exists for that workspace on the current desktop. I can imagine someone
wanting files for a workspace to be opened on a different desktop if an editor
for said workspace is already open on the other desktop, so I'll add that as a
config option later.

I don't know if I'll be able to automatically switch to that desktop (due to OS
vagaries), but even if I can't, I can at least pop up a little ephemeral window
saying that the file was opened on another desktop...