Improvements to workspace resolution.
If a file matches the hint paths of two or more workspaces, first opt for the
most recently opened workspace that's already open on the current desktop, then
fall back to the most recently opened workspace from the full list.

I also changed "opened" to mean when a workspace last had an editor created in
it, rather than when it was last resolved. Workspaces now get resolved during
the workspace search, so we don't want that to mark them as having been opened.