Bespoke collections library.
After trying to make it easier to use Scala collections from Java and other
non-Scala languages, I realized that it was simply not going to be possible to
reconcile the goal of making Scaled extensible from any JVM language and using
Scala collections throughout the Scaled APIs.

Solution: reinvent another wheel! Now Scaled has its own collections library,
which looks a lot like the Scala collections, but which does not rely on
implicits and a mashup of two dozen traits for even the most trivial
operations.

I'll document the collections and their "philosophy" in the wiki at some point,
but the executive summary is that they are (mostly) immutable and strict. If
you need mutable collections (in your API) use Java's. Internally a package can
use whatever it wants.

NOTE: the Reactual library got sucked into Scaled as a result of this because I
need Reactual to use Scaled collections (when being used in Scaled). Reactual
can live on (using Scala collections), but it was really only being used for
Scaled, so future progress will likely be confined to this "fork".